// src/types/enums/erp/flags-enum.ts
export enum FlagsTypesEnum {
  ACCOUNT_GROUP = 'Account Group',
  ACCOUNT_CLASS = 'Account Class',

  COST_CENTER_TYPE = 'Cost Center Type',
  COST_CENTER_GROUP = 'Cost Center Group',

  PROJECT_TYPE = 'Project Type',
  PROJECT_GROUP = 'Project Group',

  COVENANT_TYPE = 'Covenant Type',
  EXPENSE_TYPE = 'Expense Type',
  ASSET_TYPE = 'Asset Type',
  JURISDICTION_TYPE = 'Jurisdiction Type',
  EMPLOYEE_TYPE = 'Employee Type',

  JOURNAL_TYPE = 'Journal Type',
  CASH_RECEIT_TYPE = 'Cash Receit Type',
  BANK_RECEIT_TYPE = 'Bank Receit Type',
  CASH_PAYMENT_TYPE = 'Cash Payment Type',
  BANK_PAYMENT_TYPE = 'Bank Payment Type',
}